requirements.txt
google-adk
litellm
tavily-python
python-dotenv
yfinance
mftool
pandas
numpy
deprecated